requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Faker==9.8.0
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
